+++ R-package/tests/testthat/test-abess-group.R
@@ -38,8 +38,8 @@ test_batch <- function(abess_fit, dataset, family) {
   names(est_beta) <- NULL
   names(est_coef0) <- NULL
   
-  expect_equal(oracle_beta, est_beta, tolerance = 1e-3)
-  expect_equal(oracle_coef0, est_coef0, tolerance = 1e-3)
+  expect_equal(oracle_beta, est_beta, tolerance = 1e-2)
+  expect_equal(oracle_coef0, est_coef0, tolerance = 1e-2)
   
   ## deviance
   if (f[["family"]] == "gaussian") {
@@ -200,7 +200,7 @@ test_that("Group selection: abess (logistic) works", {
 test_that("Group selection: abess (poisson) works", {
   n <- 200
   J <- 100
-  k <- 4
+  k <- 3
   support_size <- 3
   
   dataset <-