+++ R-package/tests/testthat/test-abess.R
@@ -14,12 +14,22 @@ test_batch <- function(abess_fit, dataset, family) {
   true_index <- which(dataset[["beta"]] != 0)
   expect_equal(est_index, true_index)
   
+  
   ## estimation
   # oracle estimation by glm function:
   dat <- cbind.data.frame("y" = dataset[["y"]],
                           dataset[["x"]][, true_index])
   
-  oracle_est <- glm(y ~ ., data = dat, family = family)
+  # in gamma model, we have to set a start point of coef to ensure eta=Xb is positive  
+  if(family()[["family"]] == "Gamma"){
+    start_point <- rep(1,length(true_index))
+    coef0 <- abs(min(dataset[["x"]][, true_index] %*% start_point)) + 1
+    oracle_est <- glm(y ~ ., data = dat, family = "Gamma",start = c(coef0,start_point))
+  }
+  else{
+    oracle_est <- glm(y ~ ., data = dat, family = family)
+  }
+  
   oracle_beta <- coef(oracle_est)[-1]
   oracle_coef0 <- coef(oracle_est)[1]
   names(oracle_beta) <- NULL
@@ -43,12 +53,15 @@ test_batch <- function(abess_fit, dataset, family) {
   f <- family()
   if (f[["family"]] == "gaussian") {
     oracle_dev <- mean((oracle_est[["residuals"]]) ^ 2)
+  } else if (f[["family"]] == "Gamma") {
+    oracle_dev <- extract(abess_fit)[["dev"]]
   } else if (f[["family"]] != "poisson") {
     oracle_dev <- deviance(oracle_est) / 2
   } else {
     oracle_dev <- extract(abess_fit)[["dev"]]
   }
   expect_equal(oracle_dev, extract(abess_fit)[["dev"]])
+ 
 }
 
 test_batch_multivariate <-
@@ -309,19 +322,22 @@ test_that("abess (cox) works", {
 })
 
 test_that("abess (poisson) works", {
-  skip("poisson")
+  #skip("poisson")
   n <- 200
   p <- 100
   support.size <- 3
   
   dataset <- generate.data(n, p, support.size,
-                           family = "poisson", seed = 1)
+                           family = "poisson", seed = 78)
   abess_fit <- abess(
     dataset[["x"]],
     dataset[["y"]],
     family = "poisson",
+    #newton = "exact",
     tune.type = "cv",
-    newton.thresh = 1e-8
+    newton.thresh = 1e-8,
+    #support.size = 0:support.size,
+    #always.include = which(dataset[['beta']]!=0)
   )
   test_batch(abess_fit, dataset, poisson)
 })
@@ -446,9 +462,27 @@ test_that("abess (always-include) works", {
   p <- 20
   support_size <- 3
   dataset <- generate.data(n, p, support_size)
-  abess_fit <-
-    abess(dataset[["x"]], dataset[["y"]], always.include = c(1))
+  abess_fit <- abess(dataset[["x"]], dataset[["y"]], always.include = c(1))
   expect_true(all((abess_fit[["beta"]][1, , drop = TRUE][-1] != 0)))
+})
+
+test_that("abess (L2 regularization, ridge estimation) works", {
+  n <- 50
+  p <- 20
+  support_size <- 3
+  lambda_value <- 1
+  dataset <- generate.data(n, p, support_size)
+  true_index <- which(dataset[["beta"]] != 0)
+  abess_fit <- abess(dataset[["x"]], dataset[["y"]], always.include = true_index, 
+                     lambda = lambda_value, support.size = length(true_index))
+  coef_est <- as.vector(coef(abess_fit))
+  coef_est <- coef_est[coef_est != 0]
+  if (!require("MASS")) {
+    install.packages("MASS")
+  }
+  dat <- cbind.data.frame("y" = dataset[["y"]], "x" = dataset[["x"]][, true_index])
+  oracle_est <- coef(lm.ridge(y ~ ., data = dat, lambda = lambda_value))
+  expect_equal(as.numeric(coef_est), as.numeric(oracle_est), tolerance = 1e-10)
 })
 
 test_that("abess (L2 regularization) works", {
@@ -459,11 +493,25 @@ test_that("abess (L2 regularization) works", {
   abess_fit <- abess(dataset[["x"]], dataset[["y"]], lambda = 0.1)
   expect_true(all(diff(abess_fit[["edf"]]) > 0))
   expect_true(all(abess_fit[["edf"]] <= abess_fit[["support.size"]]))
-  
-  n <- 20
-  p <- 100
-  dataset <- generate.data(n, p, support_size)
   abess_fit2 <- abess(dataset[["x"]], dataset[["y"]])
-  expect_gt(extract(abess_fit)[["support.size"]],
-            extract(abess_fit2)[["support.size"]])
+  expect_true(extract(abess_fit)[["support.size"]] >= extract(abess_fit2)[["support.size"]])
+})
+
+test_that("abess (gamma) works", {
+  skip_on_ci()
+  n <- 2000
+  p <- 100
+  support.size <- 3
+  dataset <- generate.data(n, p, support.size,
+                           family = "gamma", seed = 1)
+  
+  abess_fit <- abess(
+    dataset[["x"]],
+    dataset[["y"]],
+    family = "gamma",
+    tune.type = "cv", 
+    support.size = 0:support.size
+  )
+  
+  test_batch(abess_fit, dataset, Gamma)
 })