+++ docs/Changelog.md
@@ -9,6 +9,12 @@
 * Python package
   * Support user-specified cross validation division by additional argument `cv_fold_id`. 
 
+* R package
+  * Support user-specified cross validation division for `abess` function by additional argument `foldid`. 
+
+* R package 
+  * Support generalized linear model when the link function is gamma distribution. It is help for the dataset with a positive valued and skewed response. 
+
 * Project development
   * Documentation
     * Add instruction for user-specified cross validation division.